New test for WSL_GUI
Right now, this capability is only being tested in SLE and from the checkbox in the firstrun. By installing the wsl_gui pattern from zypper we can include openSUSE in the tests too. Also a new test has been added to check some basic x11 functionality and not only the installation.